Extract ObservabilityConfig
